monitoring: improve diff from ES to DB calculation
* Uses timestamps to calculate exact DB ES count differences.
